requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv==1.0.0



